requirements.txt
pandas==2.2.3
kafka-python==2.0.2
clickhouse-driver==0.2.9
python-dotenv==1.0.1
